+++ tests/unittest/test_thread_client.py
@@ -196,10 +196,9 @@ def test_get_settings(fake_heksher_service, monkeypatch):
     monkeypatch.setattr(fake_heksher_service, 'settings_response', orjson.dumps(settings))
     with ThreadHeksherClient(fake_heksher_service.url, 1000, ['a', 'b', 'c']) as client:
         response = client.get_settings()
-        assert response == [
-            SettingData(name="settings",
-                        configurable_features=['a', 'b'],
-                        type='int',
-                        default_value=None,
-                        metadata={})
-        ]
+        assert response == {"settings": SettingData(name="settings",
+                                                    configurable_features=['a', 'b'],
+                                                    type='int',
+                                                    default_value=None,
+                                                    metadata={})
+                            }